added authorize attribute to postController methods and made them not require AuthorId in bodu (#24)

+++ src/Controllers/PostsController.cs
@@ -156,14 +156,20 @@ namespace Projekt_Zespolowy.Controllers
             PostDTO postDTO = result.ResponseBody;
             return result.ResponseBody != null ? Ok(result.ResponseBody) : NotFound(); 
         }
-
+        [Authorize]
         [HttpPost]
         public IActionResult Post([FromBody] PostDTO postDTO)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if(usr == null)
+            {
+                return Unauthorized();
+            }
+            postDTO.AuthorId = userManager.FindByNameAsync(usr).Result.Id;
             var response = postsService.Add(postDTO);
             if (response.ResponseCode == 201)
             {
-                return Created($"/posts/{postDTO.Id}", response.ResponseBody);
+                return Created($"/posts/{postDTO.Id}", (PostDTO)response.ResponseBody);
             }
             if (response.ResponseCode == 413)
             {
@@ -186,27 +192,40 @@ namespace Projekt_Zespolowy.Controllers
                 return StatusCode(response.ResponseCode);
             }
         }
-
+        [Authorize]
         [HttpPost("community/{community_id}")]
         public IActionResult PostInCommunity(int community_id, [FromBody] PostDTO postDTO)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (usr == null)
+            {
+                return Unauthorized();
+            }
+            postDTO.AuthorId = userManager.FindByNameAsync(usr).Result.Id;
             var response = postsService.AddInCommunity(community_id, postDTO);
-            if(response.ResponseCode == 404)
+            if (response.ResponseCode == 404)
             {
                 return NotFound("Dana społeczność nie istnieje!");
             }
             if (response.ResponseCode == 201)
             {
-                return Created($"/posts/{postDTO.Id}", response.ResponseBody);
+                return Created($"/posts/{postDTO.Id}", (PostDTO)response.ResponseBody);
             }
             else
             {
                 return StatusCode(response.ResponseCode);
             }
         }
+        [Authorize]
         [HttpPost("{parent_id}")]
         public IActionResult PostAsComment(int parent_id, [FromBody] PostDTO postDTO)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (usr == null)
+            {
+                return Unauthorized();
+            }
+            postDTO.AuthorId = userManager.FindByNameAsync(usr).Result.Id;
             var response = postsService.AddComment(parent_id, postDTO);
             if (response.ResponseCode == 404)
             {
@@ -214,16 +233,27 @@ namespace Projekt_Zespolowy.Controllers
             }
             if (response.ResponseCode == 201)
             {
-                return Created($"/posts/{postDTO.Id}", response.ResponseBody);
+                return Created($"/posts/{postDTO.Id}", (PostDTO)response.ResponseBody);
             }
             else
             {
                 return StatusCode(response.ResponseCode);
             }
         }
+        [Authorize]
         [HttpPut]
         public IActionResult Put([FromBody] PostDTO postDTO)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (usr == null)
+            {
+                return Unauthorized();
+            }
+            var usrId = userManager.FindByNameAsync(usr).Result.Id;
+            if(postsService.GetById(postDTO.Id).ResponseBody.AppUserId != usrId)
+            {
+                return Unauthorized();
+            }
             var response = postsService.Update(postDTO);
             if (response.ResponseCode == 200)
             {
@@ -234,10 +264,20 @@ namespace Projekt_Zespolowy.Controllers
                 return StatusCode(response.ResponseCode);
             }
         }
-
+        [Authorize]
         [HttpPut("[action]/{id}")]
         public IActionResult Delete(int id)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (usr == null)
+            {
+                return Unauthorized();
+            }
+            var usrId = userManager.FindByNameAsync(usr).Result.Id;
+            if (postsService.GetById(id).ResponseBody.AppUserId != usrId)
+            {
+                return Unauthorized();
+            }
             var response = postsService.Delete(id);
             if (response.ResponseCode == 200)
             {
@@ -256,9 +296,20 @@ namespace Projekt_Zespolowy.Controllers
                 return StatusCode(response.ResponseCode);
             }
         }
+        [Authorize]
         [HttpPut("[action]/{id}")]
         public IActionResult UndoDelete(int id)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (usr == null)
+            {
+                return Unauthorized();
+            }
+            var usrId = userManager.FindByNameAsync(usr).Result.Id;
+            if (postsService.GetById(id).ResponseBody.AppUserId != usrId)
+            {
+                return Unauthorized();
+            }
             var response = postsService.UndoDelete(id);
             if (response.ResponseCode == 200)
             {
@@ -277,9 +328,16 @@ namespace Projekt_Zespolowy.Controllers
                 return StatusCode(response.ResponseCode);
             }
         }
+        [Authorize]
         [HttpPost("{postId}/Like")]
         public IActionResult GiveReaction(int postId, [FromBody] LikeDTO like)
         {
+            var usr = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (usr == null)
+            {
+                return Unauthorized();
+            }
+            like.AppUserId = userManager.FindByNameAsync(usr).Result.Id;
             var post = this.postsService.GetById(postId);
             if(post.ResponseCode == StatusCodes.Status404NotFound || post.ResponseBody.IsDeleted)
             {
